Split out Kotlin DSL property assignment tests
